+++ admin/kuo.php
@@ -72,8 +72,19 @@
 											<td>Rp.<?=number_format($sisa) ?>,-</td>
 											<td><?=$b['no_bkt_pengembalian'] ?></td>
 											<td>
-												<a href="edit_kuo.php?no_bukti=<?=$no_bukti?>" class="btn btn-warning">Edit</a>
-                        <a onclick="if(confirm('Apakah anda yakin ingin menghapus data ini ??')){ location.href='hapus.php?page=kuo&id=<?=$b['id']?>' }" class="btn btn-danger">Hapus</a>
+                        <?php
+                          if($type=="admin"){
+                            ?>
+                            <a href="edit_kuo.php?no_bukti=<?=$no_bukti?>" class="btn btn-warning">Edit</a>
+                            <a onclick="if(confirm('Apakah anda yakin ingin menghapus data ini ??')){ location.href='hapus.php?page=kuo&id=<?=$b['id']?>' }" class="btn btn-danger">Hapus</a>
+                            <?php
+                          } else if($type=="editor"){
+                            ?>
+                            <a href="edit_kuo.php?no_bukti=<?=$no_bukti?>" class="btn btn-warning">Edit</a>
+                            <?php
+                          } else{
+                          }
+                        ?>
 											</td>
 										</tr>
 										<?php
