+++ admin/realisasi_spk.php
@@ -82,8 +82,20 @@
 										<td>Rp.<?=number_format($b['laba']) ?>,-</td>
 										<td><?=round($b['pm'],2) ?>%</td>
 										<td>
-										<a href="edit_real_spk.php?cus=<?=$cu?>&no_spk=<?=$id?>&id=<?=$b['id']?>" class="btn btn-warning">Edit</a>
-										<a onclick="if(confirm('Apakah anda yakin ingin menghapus data ini ??')){ location.href='hapus.php?page=realisasi_spk&cus=<?=$cu?>&no_spk=<?=$id?>&id=<?=$b['id']?>' }" class="btn btn-danger">Hapus</a>
+											<?php
+                        if($type=="admin"){
+                          ?>
+													<a href="edit_real_spk.php?cus=<?=$cu?>&no_spk=<?=$id?>&id=<?=$b['id']?>" class="btn btn-warning">Edit</a>
+													<a onclick="if(confirm('Apakah anda yakin ingin menghapus data ini ??')){ location.href='hapus.php?page=realisasi_spk&cus=<?=$cu?>&no_spk=<?=$id?>&id=<?=$b['id']?>' }" class="btn btn-danger">Hapus</a>
+                          <?php
+                			  } else if($type=="editor"){
+                          ?>
+                          <a href="edit_real_spk.php?cus=<?=$cu?>&no_spk=<?=$id?>&id=<?=$b['id']?>" class="btn btn-warning">Edit</a>
+                          <?php
+                			  } else{
+
+                			  }
+                      ?>
 										</td>
 									</tr>
 									<?php
